// src/editor-config/types.ts
export type Platform = "web" | "desktop";
export type Properties = PropertyGroup[];

type PropertyGroup = {
    caption: string;
    propertyGroups?: PropertyGroup[];
    properties?: Property[];
};

type Property = {
    key: string;
    caption: string;
    description?: string;
    objectHeaders?: string[]; // used for customizing object grids
    objects?: ObjectProperties[];
    properties?: Properties[];
};

type ObjectProperties = {
    properties: PropertyGroup[];
    captions?: string[]; // used for customizing object grids
};

export type Problem = {
    property?: string; // key of the property, at which the problem exists
    severity?: "error" | "warning" | "deprecation"; // default = "error"
    message: string; // description of the problem
    studioMessage?: string; // studio-specific message, defaults to message
    url?: string; // link with more information about the problem
    studioUrl?: string; // studio-specific link
};

/*
Standard types
Generated by xx.editorConfig.ts file and moved here as part of refactoring
*/

export type BaseProps = {
    grow?: number; // optionally sets a growth factor if used in a layout (default = 1)
};

export type ImageProps = BaseProps & {
    type: "Image";
    document?: string; // svg image
    data?: string; // base64 image
    property?: object; // widget image property object from Values API
    width?: number; // sets a fixed maximum width
    height?: number; // sets a fixed maximum height
};

export type ContainerProps = BlockStylingProps & {
    type: "Container";
    children?: PreviewProps[]; // any other preview element
};

export type RowLayoutProps = RowLayoutStylingProps & {
    type: "RowLayout";
    children: PreviewProps[];
};

export type TextProps = TextStylingProps & {
    type: "Text";
    content: string; // text that should be shown
    
};

export type DropZoneProps = DropZoneStylingProps & {
    type: "DropZone";
    property: object; // widgets property object from Values API
};

export type SelectableProps = BaseProps & {
    type: "Selectable";
    object: object; // object property instance from the Value API
    child: PreviewProps; // any type of preview property to visualize the object instance
};

export type DatasourceProps = BaseProps & {
    type: "Datasource";
    property: object | null; // datasource property object from Values API
    child?: PreviewProps; // any type of preview property component (optional)
};

export type PreviewProps =
    | ImageProps
    | ContainerProps
    | RowLayoutProps
    | TextProps
    | DropZoneProps
    | SelectableProps
    | DatasourceProps;

    /*
Custom types
Custom styling types to simplify above definitions
*/

export interface BlockStylingProps extends BaseProps {
    borders?: boolean; // sets borders around the layout to visually group its children
    borderRadius?: number; // integer. Can be used to create rounded borders
    backgroundColor?: string; // HTML color, formatted #RRGGBB
    borderWidth?: number; // sets the border width
    padding?: number; // integer. adds padding around the container
}

export interface RowLayoutStylingProps extends BlockStylingProps {
    columnSize?: "fixed" | "grow";
}

export interface TextStylingProps extends BaseProps {
    fontSize?: number;
    fontColor?: string;
    bold?: boolean;
    italic?: boolean;
}

export interface DropZoneStylingProps extends BaseProps {
    placeholder?: string; // text to be shown inside the dropzone when empty
    showDataSourceHeader?: boolean; // true by default. Toggles whether to show a header containing information about the datasource
}